requirements.txt
requests
azure-ai-formrecognizer==3.3.3
dotenv
azure-ai-inference[opentelemetry]
azure-identity
openai
tabulate
datasets
azure-ai-projects==1.1.0b3
azure-ai-agents==1.2.0b3
azure-search-documents==11.7.0b1
azure-identity
requests